Splitted source files into Core and Hal libraries
