+++ Firmware/src/Core/Readme.md
@@ -1,0 +1,3 @@
+This Core folder gathers source files that are independent from the Hardware and have no direct dependency on it.
+Several small libraries are exposed and are tested (built with x86 or whatever your CPU is, using Google Tests).
+See [Tests](Tests/) folder.
