+++ Firmware/src/Core/bridge.c
