+++ Firmware/src/Core/bridge.h
