+++ Firmware/src/Core/thermistor_ntc_100k_3950K.h
