+++ Firmware/src/Hal/Readme.md
@@ -1,0 +1,3 @@
+This folder is used to map files that use Avr code directly, and hence are more difficult to Unit test as they would require proper stubbing from the AVR world.
+
+This is a sort of HAL (Hardware Abstraction Library) and thus is really tied to avr architecture.
