+++ Firmware/src/Hal/persistent_memory.c
